refactor(reasoning/thinking/edges): Define schema, refactor DTOs
Define a dedicated Mongoose schema (ThinkingEdgeSchema) for embedded
thinking flow edge documents. Update ReasoningThinkingDocument schema
to use this schema for the edges array instead of [Object].

Remove redundant 'id' field from the ThinkingEdge interface to align
with Mongoose's automatic _id generation for subdocuments.

Move edge DTOs (ReasoningThinkingEdgeDto, CreateReasoningThinkingEdgeDto,
UpdateReasoningThinkingEdgeDto) from their separate files into the main
thinking flow DTO files (thinking.dto.ts, create-thinking.dto.ts,
update-thinking.dto.ts) for consolidation and better locality.

Update ReasoningThinkingDto constructor mapping for edges to use the
newly relocated DTO.

This improves the database modeling of edges and consolidates related DTOs
for better code organization.
