+++ src/modules/base/base.schemas.ts
@@ -24,9 +24,6 @@ export class VariableSchemaClass implements Variable {
 
   @Prop({ type: MongooseSchema.Types.Mixed })
   defaultValue?: any;
-
-  // createdAt 和 updatedAt 會由 timestamps: true 自動處理，無需在這裡使用 @Prop 定義
 }
 
-// 創建 Mongoose Schema 實例
 export const VariableSchema = SchemaFactory.createForClass(VariableSchemaClass);
