--- src/modules/reasoning/thinking/dto/update-edge.dto.ts
@@ -1,35 +1,0 @@
-import {
-  IsOptional,
-  IsString
-} from 'class-validator';
-
-export class UpdateReasoningThinkingEdgeDto {
-
-  @IsOptional()
-  @IsString()
-  source?: string;
-
-  @IsOptional()
-  @IsString()
-  sourceHandle?: string;
-
-  @IsOptional()
-  @IsString()
-  target?: string;
-
-  @IsOptional()
-  @IsString()
-  targetHandle?: string;
-
-  @IsOptional()
-  @IsString()
-  type?: string;
-
-  @IsOptional()
-  @IsString()
-  label?: string;
-
-  // Note: Fields like 'status', 'createdAt', 'updatedAt', or internal IDs (_id)
-  // are typically managed by the backend service and the database,
-  // and thus are not included in the DTO used for *creating* the resource.
-}
